+++ packages/vuetify-grid-table/src/components/grid-table/index.ts
@@ -1,12 +1,19 @@
 export { default as GridTable } from './GridTable.vue'
 export { default as GridCellEditor } from './GridCellEditor.vue'
 export type {
+  GridActiveChange,
   GridCellChange,
   GridCellRef,
   GridCellValue,
   GridColumn,
   GridColumnType,
+  GridEditEnd,
+  GridEditStart,
+  GridEvent,
   GridOption,
   GridRange,
   GridRow,
+  GridRowDelete,
+  GridRowInsert,
+  GridRowMove,
 } from './types.js'